commands/help: Extract "json-generator" utility
