+++ lib/commands/help.js
@@ -7,8 +7,7 @@ var string          = require('../utilities/string');
 var assign          = require('lodash/assign');
 var GenerateCommand = require('./generate');
 var RootCommand     = require('../utilities/root-command');
-var versionUtils    = require('../utilities/version-utils');
-var emberCLIVersion = versionUtils.emberCLIVersion;
+var JsonGenerator   = require('../utilities/json-generator');
 
 module.exports = Command.extend({
   name: 'help',
@@ -94,7 +93,14 @@ module.exports = Command.extend({
   },
 
   _printJsonHelp: function(commandOptions, rawArgs) {
-    var json = this._generateJson(commandOptions, rawArgs);
+    var generator = new JsonGenerator({
+      ui: this.ui,
+      project: this.project,
+      commands: this.commands,
+      tasks: this.tasks
+    });
+
+    var json = generator.generate(commandOptions, rawArgs);
 
     var outputJsonString = JSON.stringify(json, function(key, value) {
       // build command has a recursive property
@@ -105,73 +111,6 @@ module.exports = Command.extend({
     }, 2);
 
     this.ui.writeLine(outputJsonString);
-  },
-
-  _generateJson: function(commandOptions, rawArgs) {
-    var rootCommand = new RootCommand({
-      ui: this.ui,
-      project: this.project,
-      commands: this.commands,
-      tasks: this.tasks
-    });
-
-    var json = rootCommand.getJson(commandOptions);
-    json.version = emberCLIVersion();
-    json.commands = [];
-    json.addons = [];
-
-    if (rawArgs.length === 0) {
-      Object.keys(this.commands).forEach(function(commandName) {
-        this._addCommandHelpToJson(commandName, false, commandOptions, json);
-      }, this);
-
-      if (this.project.eachAddonCommand) {
-        this.project.eachAddonCommand(function(addonName, commands) {
-          this.commands = commands;
-
-          var addonJson = { name: addonName };
-          addonJson.commands = [];
-          json.addons.push(addonJson);
-
-          Object.keys(this.commands).forEach(function(commandName) {
-            this._addCommandHelpToJson(commandName, false, commandOptions, addonJson);
-          }, this);
-        }.bind(this));
-      }
-    } else {
-      // If args were passed to the help command,
-      // attempt to look up the command for each of them.
-
-      if (this.project.eachAddonCommand) {
-        this.project.eachAddonCommand(function(addonName, commands) {
-          assign(this.commands, commands);
-        }.bind(this));
-      }
-
-      var multipleCommands = [GenerateCommand.prototype.name].concat(GenerateCommand.prototype.aliases);
-      if (multipleCommands.indexOf(rawArgs[0]) > -1) {
-        var command = rawArgs.shift();
-        if (rawArgs.length > 0) {
-          commandOptions.rawArgs = rawArgs;
-        }
-        rawArgs = [command];
-      }
-
-      // Iterate through each arg beyond the initial 'help' command,
-      // and try to display usage instructions.
-      rawArgs.forEach(function(commandName) {
-        this._addCommandHelpToJson(commandName, true, commandOptions, json);
-      }, this);
-    }
-
-    return json;
-  },
-
-  _addCommandHelpToJson: function(commandName, single, options, json) {
-    var command = this._lookupCommand(commandName);
-    if (!command.skipHelp || single) {
-      json.commands.push(command.getJson(options));
-    }
   },
 
   _printHelpForCommand: function(commandName, detailed, options) {
